+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Change Log
 ==========
 
+Version 2.0.1 *(11 August, 2016)*
+-------------------------------------------
+*(Supports analytics-ios 3.3+ and Primer 3.1+)*
+
+  * Improved Primer event forwarding.
+
 Version 2.0.0 *(10 August, 2016)*
 -------------------------------------------
 *(Supports analytics-ios 3.3+ and Primer 3.1+)*